Convert package config from coffee to javascript
For now it appears the package config file must be in javascript. It's
in the config file that we declare our dependency on coffeescript, and
so Meteor ignores the coffee config file, and therefore none of the
package is loaded. We've been getting by with auto-transpiled files in
Webstorm, but it's almost time for a real package release.

Update gitignore to exclude the package.js file from the set of ignored
files (all js files otherwise ignored).