Fix Issue #1
The default api path ('api/') is now used if no apiPath is provided in
Restivus.configure().

A separate issue was also discovered and resolved in Restivus.configure()
where an apiPath ending with '/' was having an extra '/' appended to it.
Fix so it only happens when the path is missing the trailing '/'.