Fix Issue #4
Resolve and configure endpoints in Route.addToApi() instead of Route
constructor so that endpoints are never configured until the API has
been configured (with Restivus.configure()). This was causing unexpected
behavior in routes that were added before the API was configured, such
as authentication not being checked when necessary (so this.user and
this.userId were undefined in authenticated endpoints).