Allow existing collections to be added to API
Fix Issue #10:
v0.6.0 does not allow you to work with existing collections, or access
collections created in Restivus elsewhere in your app, which is not the
intended behavior. Since Meteor expects you to construct each collection
only once (using `new Mongo.Collection()`), and store that globally,
Restivus now requires that you pass an existing collection in
`Restivus.addCollection()`, in place of the name. Please check out the
updated section of the README on Defining Collection Routes for a
detailed breakdown of the parameters to `Restivus.addCollection()`.